Machine Learning with Cloud Computing/Practical 3/requirements.txt
cryptography==42.0.3 ## 'cryptography' package is required for sha256_password or caching_sha2_password auth method
pandas==2.2.0
pyarrow==15.0.0 ## Pyarrow will become a required dependency of pandas in the next major release of pandas (pandas 3.0)
pymysql==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.27
